Navigation finished, app icon added

+++ app/src/main/java/com/iesnervion/keepitfitness/domain/model/EjercicioEntrenamiento.kt
@@ -2,6 +2,6 @@ package com.iesnervion.keepitfitness.domain.model
 
 data class EjercicioEntrenamiento(
     var excersice: Ejercicio,
-    var reps: Int,
-    var weight: Int
+    var reps: Long,
+    var weight: Long
 )
